fix:Rename ddns.fm to LightDNS
